Fixed Firefox by allowing recovery after ondataavailable processing fails.
It appears that Firefox's ogg decoder was failing if too short a buffer
was passed to it. Because the error handling was broken that led to a
permanent failure. Now the error is caught, and the next ondataavailable
callback will be able to retry the processing, which will succeed once
enough data has been recorded.