Add playlist container
Tracks do not render, because the console complains "cannot appendChild to null". Adding `<div id="playlist"></div>` solves the problem.
However, being new to this project, I do not know if there is a build process required (and if, if it is automated) or not (and I therefore should add the container somewhere else, too) to add this container to the `dist/waveform-playlist`.